requirements.txt
django==2.1.1
mysqlclient==1.3.12
honcho==1.0.1
gunicorn==19.8.1
WhiteNoise==3.3.1
telepot
google-api-python-client==1.7.4
python-dateutil
oauth2client==4.1.3
djangorestframework
markdown
django-filter
djangorestframework-jwt
coreapi
django-rest-swagger
sentry_sdk
django-q
arrow